Trekk ut JournalpostDokument til egen komponent
